# INSTA-LITE/insta-lite-back/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginRequest:
    properties:
      identifier:
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: hugues@gmail.com
        type: string
      fullName:
        example: Goli yao Hugues
        type: string
      password:
        example: pass1234
        minLength: 6
        type: string
      username:
        example: huguesgoli
        minLength: 3
        type: string
    required:
    - email
    - fullName
    - password
    - username
    type: object
  comment.Comment:
    properties:
      commentId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      likesCount:
        type: integer
      parentCommentId:
        type: string
      postId:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  comment.CommentCreateRequest:
    properties:
      content:
        maxLength: 2000
        type: string
      parentCommentId:
        type: string
    required:
    - content
    type: object
  comment.CommentUpdateRequest:
    properties:
      content:
        maxLength: 2000
        type: string
    required:
    - content
    type: object
  comment.PaginatedCommentResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/comment.Comment'
        type: array
      hasMore:
        type: boolean
      nextCursor:
        type: string
    type: object
  commentlike.CommentLike:
    properties:
      commentId:
        type: string
      createdAt:
        type: string
      reactionType:
        $ref: '#/definitions/like.ReactionType'
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  commentlike.reactionRequest:
    properties:
      reactionType:
        $ref: '#/definitions/like.ReactionType'
    required:
    - reactionType
    type: object
  follow.Follow:
    properties:
      createdAt:
        type: string
      followeeId:
        type: string
      followerId:
        type: string
      status:
        $ref: '#/definitions/follow.FollowStatus'
    type: object
  follow.FollowStatus:
    enum:
    - pending
    - accepted
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusAccepted
...
  like.ReactionType:
    enum:
    - like
    - love
    - laugh
    type: string
    x-enum-varnames:
    - ReactionLike
    - ReactionLove
    - ReactionLaugh
  notification.Notification:
    properties:
      actorId:
        type: string
      createdAt:
        type: string
      entityId:
        type: string
      notificationId:
        type: string
      read:
        type: boolean
      recipientId:
        type: string
      type:
        $ref: '#/definitions/notification.NotificationType'
    type: object
  notification.NotificationType:
    enum:
    - like_post
    - comment_post
    - like_comment
    - new_follower
    type: string
    x-enum-varnames:
    - TypeLikePost
    - TypeCommentPost
    - TypeLikeComment
    - TypeNewFollower
  notification.PaginatedNotificationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/notification.Notification'
        type: array
      hasMore:
        type: boolean
      nextCursor:
        type: string
      unreadCount:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: huguescodeur@gmail.com
    name: Hugues Codeur
  description: API REST pour une application de type Instagram (posts, likes, commentaires,
    follow).
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Insta-Lite API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authentifie un utilisateur (email ou username) et retourne le token
        JWT
      parameters:
      - description: Identifiants de connexion
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format JSON invalide ou validation échouée
          schema:
            type: string
        "401":
          description: Identifiants incorrects
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      summary: Connexion
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Crée un compte et retourne le token JWT
      parameters:
      - description: Données d'inscription
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format JSON invalide ou validation échouée
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      summary: Inscription d'un nouvel utilisateur
      tags:
      - Auth
  /comments/{commentID}/likes:
    delete:
      description: Retire la réaction de l'utilisateur connecté sur un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      responses:
        "204":
          description: Like supprimé
        "400":
          description: ID incorrect
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Supprimer un like sur un commentaire
      tags:
      - Comment Likes
    get:
      description: Retourne tous les likes d'un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/commentlike.CommentLike'
            type: array
        "400":
          description: ID incorrect
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Likes d'un commentaire
      tags:
      - Comment Likes
    patch:
      consumes:
      - application/json
      description: Change le type de réaction de l'utilisateur connecté sur un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      - description: Nouvelle réaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commentlike.reactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commentlike.CommentLike'
        "400":
          description: Type invalide
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Modifier une réaction sur un commentaire
      tags:
      - Comment Likes
    post:
      consumes:
      - application/json
      description: Ajoute une réaction (like/love/laugh) à un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      - description: Type de réaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commentlike.reactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/commentlike.CommentLike'
        "400":
          description: Type de réaction invalide
          schema:
            type: string
        "409":
          description: Déjà liké
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Liker un commentaire
      tags:
      - Comment Likes
  /comments/{commentID}/likes/me:
    get:
      description: Retourne le like de l'utilisateur connecté sur un commentaire
      parameters:
      - description: UUID du commentaire
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commentlike.CommentLike'
        "400":
          description: ID incorrect
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Mon like sur un commentaire
      tags:
      - Comment Likes
  /comments/{id}:
    delete:
      description: Supprime un commentaire appartenant à l'utilisateur connecté
      parameters:
      - description: UUID du commentaire
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Commentaire supprimé
        "400":
          description: ID incorrect
          schema:
            type: string
        "404":
          description: Commentaire non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Supprimer un commentaire
      tags:
      - Comments
    patch:
      consumes:
      - application/json
      description: Met à jour le contenu d'un commentaire appartenant à l'utilisateur
        connecté
      parameters:
      - description: UUID du commentaire
        in: path
        name: id
        required: true
        type: string
      - description: Nouveau contenu
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/comment.CommentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.Comment'
        "400":
          description: Format invalide
          schema:
            type: string
        "404":
          description: Commentaire non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Modifier un commentaire
      tags:
      - Comments
  /comments/{id}/replies:
    get:
      description: Retourne toutes les réponses d'un commentaire parent
      parameters:
      - description: UUID du commentaire parent
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/comment.Comment'
            type: array
        "400":
          description: ID incorrect
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Réponses à un commentaire
      tags:
      - Comments
...
  /notifications:
    get:
      description: Returns the authenticated user's notifications in reverse chronological
        order with cursor-based pagination and an unread count
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: 'Number of results (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.PaginatedNotificationResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - Notifications
  /notifications/{id}/read:
    patch:
      description: Marks a single notification as read. Only the recipient can mark
        their own notification.
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No content
        "400":
          description: Invalid notification ID
          schema:
            type: string
        "404":
          description: Notification not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - Notifications
...
  /posts/{postID}/comments:
    get:
      description: Retourne les commentaires racines d'un post, paginés par curseur
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      - description: Curseur de pagination
        in: query
        name: cursor
        type: string
      - description: 'Nombre de résultats (défaut: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.PaginatedCommentResponse'
        "400":
          description: ID incorrect
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Commentaires d'un post
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Ajoute un commentaire (ou une réponse) sur un post. Passer parentCommentId
        pour une réponse.
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      - description: Contenu du commentaire
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/comment.CommentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/comment.Comment'
        "400":
          description: Format invalide ou commentaire parent ne correspond pas au
            post
          schema:
            type: string
        "404":
          description: Post ou commentaire parent non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Créer un commentaire
      tags:
      - Comments
...
  /users/{userID}/follow:
    delete:
      description: L'utilisateur connecté arrête de suivre un autre utilisateur
      parameters:
      - description: UUID de l'utilisateur à ne plus suivre
        in: path
        name: userID
        required: true
        type: string
      responses:
        "204":
          description: Unfollow réussi
        "400":
          description: ID incorrect
          schema:
            type: string
        "404":
          description: Relation de follow non trouvée
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Ne plus suivre un utilisateur
      tags:
      - Follow
    post:
      description: L'utilisateur connecté commence à suivre un autre utilisateur
      parameters:
      - description: UUID de l'utilisateur à suivre
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/follow.Follow'
        "400":
          description: ID incorrect ou tentative de se suivre soi-même
          schema:
            type: string
        "409":
          description: Déjà en train de suivre
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Suivre un utilisateur
      tags:
      - Follow
  /users/{userID}/follow/status:
    get:
      description: Vérifie si l'utilisateur connecté suit un autre utilisateur (pending/accepted)
      parameters:
      - description: UUID de l'utilisateur cible
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/follow.Follow'
        "400":
          description: ID incorrect
          schema:
            type: string
        "404":
          description: Relation non trouvée
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Statut de follow
      tags:
      - Follow
  /users/{userID}/followers:
    get:
      description: Retourne la liste paginée des followers d'un utilisateur
      parameters:
      - description: UUID de l'utilisateur
        in: path
        name: userID
        required: true
        type: string
      - description: 'Numéro de page (défaut: 1)'
        in: query
        name: page
        type: integer
      - description: 'Nombre de résultats (défaut: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID incorrect
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Followers d'un utilisateur
      tags:
      - Follow
  /users/{userID}/following:
    get:
      description: Retourne la liste paginée des utilisateurs que suit un utilisateur
      parameters:
      - description: UUID de l'utilisateur
        in: path
        name: userID
        required: true
        type: string
      - description: 'Numéro de page (défaut: 1)'
        in: query
        name: page
        type: integer
      - description: 'Nombre de résultats (défaut: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID incorrect
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Abonnements d'un utilisateur
      tags:
      - Follow
securityDefinitions:
  BearerAuth:
    description: 'Entrez le token JWT avec le préfixe "Bearer ". Exemple: "Bearer
      eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"